프로그래머스 2022 KAKAO BLIND RECRUITMENT k진수에서 소수 개수 구하기 - 풀이 완료
isPrime 메서드 안에서 str을 숫자로 parsing 할 때 int 대신에 long을 사용해서 해결

+++ Solution.java
@@ -1,6 +1,7 @@
 class Solution {
     public int solution(int n, int k) {
         int answer = 0;
+
         String radixString = Integer.toString(n, k);
         String[] strArr = radixString.split("0");
 
@@ -15,7 +16,8 @@ class Solution {
     }
 
     private boolean isPrime(String str) {
-        int intVal = Integer.parseInt(str);
+        long intVal = Long.parseLong(str);
+        
         for (int i = 2; i <= (int) Math.pow(intVal, 0.5); i++) {
             if (intVal % i == 0) {
                 return false;
